Rebase and comments
